+++ CHANGELOG.md
@@ -4,3 +4,5 @@
 * extended Traceback to fetch current stack frame if no exception is in progress and no stack frame is given
 * added a feature to load a Traceback from a stream
 * added __eq__ methods to Traceback classes
+* fixed a bug with unserializing a Traceback from JSON
+